One shape per concept for state hoisting
Eleven ways to say "a value and a way to change it" collapse to five:
value/onValueChange, checked/onCheckedChange, selected/onSelectedChange,
expanded/onExpandedChange, visible/onDismissRequest.

  onSelect            -> onSelectedChange   RadioGroup, SegmentedControl,
                                            WheelPicker, DatePicker,
                                            DateRangePicker, CalendarMonth,
                                            ColorSwatchPicker
  expanded            -> visible            Popover, DropdownMenu,
                                            AnchoredDropdownMenu — Dialog and
                                            ModalBottomSheet already said
                                            visible for the same idea
  onValuesChange      -> onValueChange      MultiSelect
  page/onPageChange   -> value/onValueChange  Pagination
  time/onTimeChange   -> value/onValueChange  TimePicker, TimeField
  rememberToasts      -> rememberToastHostState
  rememberConfirmations -> rememberConfirmationController

SelectionRow and IconToggleButton take a nullable callback now, matching
Checkbox, RadioButton and Switch. Passing `null` means "the parent owns
the press, you show the state" — the shape every settings screen ends up
in. Both needed the semantics branch that goes with it, or an inert
control announces as a name with no state; InertControlPublishesStateTest
grew two cases and both fail when either branch is reverted.

Rating emitted Int and took Float, so every call site converted in one
direction and back in the other. It emits Float now. The fraction is not
decoration — an average shows a half mark, a tap cannot produce one — so
the parameter stays Float and the callback matches it. The catalog's
`score.value.toFloat()` is gone.

Three of the ten the survey ranked turn out to be right as they are, and
are left alone rather than renamed into a worse shape:

TriStateCheckbox.onClick is not Checkbox.onCheckedChange missing an
argument. It cycles rather than sets, and the caller owns the cycle;
`onClick` already means "pressed, you decide" everywhere else in the
library. Renaming it to onCheckedChange while it still receives nothing
would make it the only onXChange that does not carry the new value.

BottomSheet.state is required and ModalBottomSheet.state is defaulted
because they are driven differently: the modal one takes `visible`, and
the non-modal one has no other handle — callers read visibleHeight off
that same state to inset a map. Defaulting it would leave a sheet nobody
can open.

SearchField.onQueryChange has no hoisted `query` beside it. It is a
debounced notification sitting next to onSearch, not half of a state
pair, and calling it onValueChange would promise it fires on every
keystroke when it fires 250ms later.

No golden moved, which is the rule for this stage: renaming parameters
changes what compiles, never what draws.

+++ ui/src/commonMain/kotlin/io/kontour/ui/components/selection/Rating.kt
@@ -59,7 +59,10 @@ import kotlin.math.ceil
  * @param value The score. Clamped into `0f..count`.
  * @param contentDescription What the score is *of* — "Your rating", "Average
  *   rating". Required: five stars with no name is a row of decoration.
- * @param onValueChange `null` for read-only. Receives whole numbers only.
+ * @param onValueChange `null` for read-only. Receives whole numbers, as the
+ *   `Float` [value] is, so the emitted score assigns straight back without a
+ *   conversion at every call site. A tap cannot produce a half mark; an average
+ *   can, which is why the parameter is a `Float` in the first place.
  * @param icon The empty mark. Defaults to a star outline.
  * @param filledIcon The full mark. Defaults to a filled star.
  */
@@ -69,7 +72,7 @@ fun Rating(
     contentDescription: String,
     modifier: Modifier = Modifier,
     enabled: Boolean = true,
-    onValueChange: ((Int) -> Unit)? = null,
+    onValueChange: ((Float) -> Unit)? = null,
     count: Int = RatingDefaults.Count,
     icon: ImageVector? = null,
     filledIcon: ImageVector? = null,
@@ -140,7 +143,7 @@ fun Rating(
                         selected = markValue <= selected,
                         onClick = {
                             feedback.perform(FeedbackIntent.Selection)
-                            onValueChange(markValue)
+                            onValueChange(markValue.toFloat())
                         },
                         enabled = enabled,
                         role = androidx.compose.ui.semantics.Role.RadioButton,
